move Secrets Codec impl to codec.rs
